Add error highlighting on invalid email submission
Server side email error responses are limited to the recipient email
being invalid so I put the error message onto that by default.

Technically there is also a request method error, but that would only
happen with someone tampering with the html or javascript, so at that
point, they get what they deserve.